VenuesCarousel Component
Built a new component to take in an array of businesses and render
a horizontal scroll box with scroll snap functionality. The scroll
box only displays one business at a time. This component requires an
'id' prop and an 'items' prop for passing in a unique id and a list
of businesses to render, respectively.